perlin noise is an example, and doesnt belong as part of package... renamed scripts -> example accordingly
